Moved all onBlur helpers into reducers/formHelpers.js
